Move logic into shared helper
